chore: rename examples, add tsconfig file to prevent type lib checking in examples
