+++ examples/lazy-load/tsconfig.json
