+++ examples/plugin-resize/tsconfig.json
