+++ examples/react-gondel/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gondel-react-example",
+  "name": "gondel-react-gondel-example",
   "private": true,
   "version": "0.0.1",
   "scripts": {
