+++ examples/render-jsx/babel.config.js
