+++ examples/render-jsx/package.json
