+++ examples/render-jsx/src/components/App.tsx
