+++ examples/render-jsx/src/index.ts
